Move ZTSClient.h
